requirements.txt
Flask==3.1.0Flask-SQLAlchemy==3.0.3psycopg2-binary==2.9.7python-dotenv==1.0.0  